+++ src/Client.php
@@ -1,21 +1,21 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Http\Client\Curl;
 
 use Http\Client\Exception;
 use Http\Client\HttpAsyncClient;
 use Http\Client\HttpClient;
-use Http\Discovery\MessageFactoryDiscovery;
-use Http\Discovery\StreamFactoryDiscovery;
-use Http\Message\MessageFactory;
-use Http\Message\StreamFactory;
-use Http\Promise\Promise;
+use Http\Discovery\Psr17FactoryDiscovery;
 use Psr\Http\Message\RequestInterface;
+use Psr\Http\Message\ResponseFactoryInterface;
 use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\StreamFactoryInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
 /**
- * PSR-7 compatible cURL based HTTP client.
+ * PSR-18 and HTTPlug Async client based on lib-curl.
  *
  * @license http://opensource.org/licenses/MIT MIT
  * @author  Михаил Красильников <m.krasilnikov@yandex.ru>
@@ -32,19 +32,19 @@ class Client implements HttpClient, HttpAsyncClient
      *
      * @var array
      */
-    private $options;
+    private $curlOptions;
 
     /**
-     * PSR-7 message factory.
+     * PSR-17 response factory.
      *
-     * @var MessageFactory
+     * @var ResponseFactoryInterface
      */
-    private $messageFactory;
+    private $responseFactory;
 
     /**
-     * PSR-7 stream factory.
+     * PSR-17 stream factory.
      *
-     * @var StreamFactory
+     * @var StreamFactoryInterface
      */
     private $streamFactory;
 
@@ -53,37 +53,37 @@ class Client implements HttpClient, HttpAsyncClient
      *
      * @var resource|null
      */
-    private $handle = null;
+    private $handle;
 
     /**
      * Simultaneous requests runner.
      *
      * @var MultiRunner|null
      */
-    private $multiRunner = null;
+    private $multiRunner;
 
     /**
-     * @param MessageFactory|null $messageFactory HTTP Message factory
-     * @param StreamFactory|null  $streamFactory  HTTP Stream factory
-     * @param array               $options        cURL options {@link http://php.net/curl_setopt}
+     * @param ResponseFactoryInterface|null $responseFactory PSR-17 HTTP response factory.
+     * @param StreamFactoryInterface|null   $streamFactory   PSR-17 HTTP stream factory.
+     * @param array                         $options         cURL options {@link http://php.net/curl_setopt}
      *
      * @throws \Http\Discovery\Exception\NotFoundException If factory discovery failed
-     *
-     * @since 1.0
      */
     public function __construct(
-        MessageFactory $messageFactory = null,
-        StreamFactory $streamFactory = null,
+        ResponseFactoryInterface $responseFactory = null,
+        StreamFactoryInterface $streamFactory = null,
         array $options = []
     ) {
-        $this->messageFactory = $messageFactory ?: MessageFactoryDiscovery::find();
-        $this->streamFactory = $streamFactory ?: StreamFactoryDiscovery::find();
+        $this->responseFactory = $responseFactory ?: Psr17FactoryDiscovery::findResponseFactory();
+        $this->streamFactory = $streamFactory ?: Psr17FactoryDiscovery::findStreamFactory();
         $resolver = new OptionsResolver();
-        $resolver->setDefaults([
-            CURLOPT_HEADER => false,
-            CURLOPT_RETURNTRANSFER => false,
-            CURLOPT_FOLLOWLOCATION => false,
-        ]);
+        $resolver->setDefaults(
+            [
+                CURLOPT_HEADER => false,
+                CURLOPT_RETURNTRANSFER => false,
+                CURLOPT_FOLLOWLOCATION => false,
+            ]
+        );
         $resolver->setAllowedValues(CURLOPT_HEADER, [false]); // our parsing will fail if this is set to true
         $resolver->setAllowedValues(CURLOPT_RETURNTRANSFER, [false]); // our parsing will fail if this is set to true
 
@@ -91,7 +91,7 @@ class Client implements HttpClient, HttpAsyncClient
         // Make sure that we accept everything that is in the options.
         $resolver->setDefined(array_keys($options));
 
-        $this->options = $resolver->resolve($options);
+        $this->curlOptions = $resolver->resolve($options);
     }
 
     /**
@@ -105,11 +105,7 @@ class Client implements HttpClient, HttpAsyncClient
     }
 
     /**
-     * Sends a PSR-7 request.
-     *
-     * @param RequestInterface $request
-     *
-     * @return ResponseInterface
+     * {@inheritdoc}
      *
      * @throws \Http\Client\Exception\NetworkException In case of network problems
      * @throws \Http\Client\Exception\RequestException On invalid request
@@ -156,11 +152,7 @@ class Client implements HttpClient, HttpAsyncClient
     }
 
     /**
-     * Sends a PSR-7 request in an asynchronous way.
-     *
-     * @param RequestInterface $request
-     *
-     * @return Promise
+     * {@inheritdoc}
      *
      * @throws \Http\Client\Exception\RequestException On invalid request
      * @throws \InvalidArgumentException               For invalid header names or values
@@ -190,36 +182,31 @@ class Client implements HttpClient, HttpAsyncClient
     /**
      * Update cURL options for this request and hook in the response builder.
      *
-     * @param RequestInterface $request
-     * @param ResponseBuilder  $responseBuilder
-     *
      * @throws \Http\Client\Exception\RequestException On invalid request
      * @throws \InvalidArgumentException               For invalid header names or values
      * @throws \RuntimeException                       If can not read body
-     *
-     * @return array
      */
-    private function prepareRequestOptions(RequestInterface $request, ResponseBuilder $responseBuilder)
+    private function prepareRequestOptions(RequestInterface $request, ResponseBuilder $responseBuilder): array
     {
-        $options = $this->options;
+        $curlOptions = $this->curlOptions;
 
         try {
-            $options[CURLOPT_HTTP_VERSION]
+            $curlOptions[CURLOPT_HTTP_VERSION]
                 = $this->getProtocolVersion($request->getProtocolVersion());
         } catch (\UnexpectedValueException $e) {
             throw new Exception\RequestException($e->getMessage(), $request);
         }
-        $options[CURLOPT_URL] = (string) $request->getUri();
+        $curlOptions[CURLOPT_URL] = (string) $request->getUri();
 
-        $options = $this->addRequestBodyOptions($request, $options);
+        $curlOptions = $this->addRequestBodyOptions($request, $curlOptions);
 
-        $options[CURLOPT_HTTPHEADER] = $this->createHeaders($request, $options);
+        $curlOptions[CURLOPT_HTTPHEADER] = $this->createHeaders($request, $curlOptions);
 
         if ($request->getUri()->getUserInfo()) {
-            $options[CURLOPT_USERPWD] = $request->getUri()->getUserInfo();
+            $curlOptions[CURLOPT_USERPWD] = $request->getUri()->getUserInfo();
         }
 
-        $options[CURLOPT_HEADERFUNCTION] = function ($ch, $data) use ($responseBuilder) {
+        $curlOptions[CURLOPT_HEADERFUNCTION] = function ($ch, $data) use ($responseBuilder) {
             $str = trim($data);
             if ('' !== $str) {
                 if (strpos(strtolower($str), 'http/') === 0) {
@@ -232,21 +219,17 @@ class Client implements HttpClient, HttpAsyncClient
             return strlen($data);
         };
 
-        $options[CURLOPT_WRITEFUNCTION] = function ($ch, $data) use ($responseBuilder) {
+        $curlOptions[CURLOPT_WRITEFUNCTION] = function ($ch, $data) use ($responseBuilder) {
             return $responseBuilder->getResponse()->getBody()->write($data);
         };
 
-        return $options;
+        return $curlOptions;
     }
 
     /**
      * Return cURL constant for specified HTTP version.
      *
-     * @param string $requestVersion
-     *
      * @throws \UnexpectedValueException If unsupported version requested
-     *
-     * @return int
      */
     private function getProtocolVersion(string $requestVersion): int
     {
@@ -267,13 +250,8 @@ class Client implements HttpClient, HttpAsyncClient
 
     /**
      * Add request body related cURL options.
-     *
-     * @param RequestInterface $request
-     * @param array            $options
-     *
-     * @return array
      */
-    private function addRequestBodyOptions(RequestInterface $request, array $options): array
+    private function addRequestBodyOptions(RequestInterface $request, array $curlOptions): array
     {
         /*
          * Some HTTP methods cannot have payload:
@@ -294,40 +272,37 @@ class Client implements HttpClient, HttpAsyncClient
                 // Message has non empty body.
                 if (null === $bodySize || $bodySize > 1024 * 1024) {
                     // Avoid full loading large or unknown size body into memory
-                    $options[CURLOPT_UPLOAD] = true;
+                    $curlOptions[CURLOPT_UPLOAD] = true;
                     if (null !== $bodySize) {
-                        $options[CURLOPT_INFILESIZE] = $bodySize;
+                        $curlOptions[CURLOPT_INFILESIZE] = $bodySize;
                     }
-                    $options[CURLOPT_READFUNCTION] = function ($ch, $fd, $length) use ($body) {
+                    $curlOptions[CURLOPT_READFUNCTION] = function ($ch, $fd, $length) use ($body) {
                         return $body->read($length);
                     };
                 } else {
                     // Small body can be loaded into memory
-                    $options[CURLOPT_POSTFIELDS] = (string) $body;
+                    $curlOptions[CURLOPT_POSTFIELDS] = (string) $body;
                 }
             }
         }
 
         if ($request->getMethod() === 'HEAD') {
             // This will set HTTP method to "HEAD".
-            $options[CURLOPT_NOBODY] = true;
+            $curlOptions[CURLOPT_NOBODY] = true;
         } elseif ($request->getMethod() !== 'GET') {
             // GET is a default method. Other methods should be specified explicitly.
-            $options[CURLOPT_CUSTOMREQUEST] = $request->getMethod();
+            $curlOptions[CURLOPT_CUSTOMREQUEST] = $request->getMethod();
         }
 
-        return $options;
+        return $curlOptions;
     }
 
     /**
      * Create headers array for CURLOPT_HTTPHEADER.
      *
-     * @param RequestInterface $request
-     * @param array            $options cURL options
-     *
      * @return string[]
      */
-    private function createHeaders(RequestInterface $request, array $options): array
+    private function createHeaders(RequestInterface $request, array $curlOptions): array
     {
         $curlHeaders = [];
         $headers = $request->getHeaders();
@@ -338,10 +313,10 @@ class Client implements HttpClient, HttpAsyncClient
                 continue;
             }
             if ('content-length' === $header) {
-                if (array_key_exists(CURLOPT_POSTFIELDS, $options)) {
+                if (array_key_exists(CURLOPT_POSTFIELDS, $curlOptions)) {
                     // Small body content length can be calculated here.
-                    $values = [strlen($options[CURLOPT_POSTFIELDS])];
-                } elseif (!array_key_exists(CURLOPT_READFUNCTION, $options)) {
+                    $values = [strlen($curlOptions[CURLOPT_POSTFIELDS])];
+                } elseif (!array_key_exists(CURLOPT_READFUNCTION, $curlOptions)) {
                     // Else if there is no body, forcing "Content-length" to 0
                     $values = [0];
                 }
@@ -359,21 +334,13 @@ class Client implements HttpClient, HttpAsyncClient
         return $curlHeaders;
     }
 
-    /**
-     * Create new ResponseBuilder instance.
-     *
-     * @return ResponseBuilder
-     *
-     * @throws \RuntimeException If creating the stream from $body fails
-     */
     private function createResponseBuilder(): ResponseBuilder
     {
-        try {
-            $body = $this->streamFactory->createStream(fopen('php://temp', 'w+b'));
-        } catch (\InvalidArgumentException $e) {
-            throw new \RuntimeException('Can not create "php://temp" stream.');
-        }
-        $response = $this->messageFactory->createResponse(200, null, [], $body);
+        $body = $this->streamFactory->createStreamFromFile('php://temp', 'w+b');
+
+        $response = $this->responseFactory
+            ->createResponse(200)
+            ->withBody($body);
 
         return new ResponseBuilder($response);
     }