+++ src/MultiRunner.php
@@ -19,7 +19,7 @@ class MultiRunner
      *
      * @var resource|null
      */
-    private $multiHandle = null;
+    private $multiHandle;
 
     /**
      * Awaiting cores.
@@ -43,7 +43,7 @@ class MultiRunner
      *
      * @param PromiseCore $core
      */
-    public function add(PromiseCore $core)
+    public function add(PromiseCore $core): void
     {
         foreach ($this->cores as $existed) {
             if ($existed === $core) {
@@ -64,7 +64,7 @@ class MultiRunner
      *
      * @param PromiseCore $core
      */
-    public function remove(PromiseCore $core)
+    public function remove(PromiseCore $core): void
     {
         foreach ($this->cores as $index => $existed) {
             if ($existed === $core) {
@@ -81,7 +81,7 @@ class MultiRunner
      *
      * @param PromiseCore|null $targetCore
      */
-    public function wait(PromiseCore $targetCore = null)
+    public function wait(PromiseCore $targetCore = null): void
     {
         do {
             $status = curl_multi_exec($this->multiHandle, $active);
@@ -118,7 +118,7 @@ class MultiRunner
      *
      * @return PromiseCore|null
      */
-    private function findCoreByHandle($handle)
+    private function findCoreByHandle($handle): ?PromiseCore
     {
         foreach ($this->cores as $core) {
             if ($core->getHandle() === $handle) {