+++ tests/Functional/HttpAsyncClientGuzzleTest.php
@@ -2,22 +2,22 @@
 
 declare(strict_types=1);
 
-namespace Http\Client\Curl\Tests;
+namespace Http\Client\Curl\Tests\Functional;
 
 use Http\Client\Curl\Client;
-use Http\Client\HttpClient;
+use Http\Client\HttpAsyncClient;
 use Http\Message\MessageFactory\GuzzleMessageFactory;
 use Http\Message\StreamFactory\GuzzleStreamFactory;
 
 /**
- * Tests for Http\Client\Curl\Client.
+ * @covers \Http\Client\Curl\Client
  */
 class HttpAsyncClientGuzzleTest extends HttpAsyncClientTestCase
 {
     /**
-     * @return HttpClient
+     * {@inheritdoc}
      */
-    protected function createHttpAsyncClient()
+    protected function createHttpAsyncClient(): HttpAsyncClient
     {
         return new Client(new GuzzleMessageFactory(), new GuzzleStreamFactory());
     }